+++ routes/web.php
@@ -346,6 +346,7 @@ Route::middleware(['admin', 'nocache'])->group(function () {
     Route::get('/admin/workshops/create', [WorkshopController::class, 'admin_create'])->name('admin.workshop.create');
     Route::get('/admin/workshops/{workshop}/duplicate', [WorkshopController::class, 'admin_duplicate'])->name('admin.workshop.duplicate');
     Route::get('/admin/workshops/{workshop}/tickets', [WorkshopController::class, 'admin_tickets'])->name('admin.workshop.tickets');
+    Route::post('/admin/workshops/{workshop}/tickets', [WorkshopController::class, 'admin_tickets_store'])->name('admin.workshop.tickets.store');
     Route::get('/admin/workshops/{workshop}/tickets/pdf', [WorkshopController::class, 'admin_tickets_pdf'])->name('admin.workshop.tickets.pdf');
     Route::post('/admin/workshops/{workshop}/tickets/email', [WorkshopController::class, 'admin_tickets_email'])->name('admin.workshop.tickets.email');
     Route::get('/admin/workshops/{workshop}/attendance', [WorkshopController::class, 'admin_attendance'])->name('admin.workshop.attendance');